// ui/src/generated/models/environmentType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Lintel
 * OpenAPI spec version: 0.1.0
 */

export type EnvironmentType = typeof EnvironmentType[keyof typeof EnvironmentType];


export const EnvironmentType = {
  development: 'development',
  staging: 'staging',
  production: 'production',
  sandbox: 'sandbox',
} as const;